fix: 쓸데없는 코드 수정
