+++ app/asset/api.tsx
@@ -1,3 +1,3 @@
-const api = 'https://api.themoviedb.org/3';
+const api = "https://api.themoviedb.org/3/movie";
 
 export default api;
