+++ app/pages/_app.tsx
@@ -1,10 +1,17 @@
 import React from "react";
 import "../styles/globals.css";
+import Head from "next/head";
+import Navigator from "../component/navigation/footer";
+import styles from "../styles/Home.module.css";
 
 const App = ({ Component, pageProps }: any) => {
   return (
     <>
-      <Component {...pageProps}/>
+      <Head>
+        <title>Netfilx</title>
+        <link rel="icon" href="https://ifh.cc/g/bCkXTX.png" />
+      </Head>
+      <Component {...pageProps} />
     </>
   );
 };
