+++ app/pages/index.tsx
@@ -1,6 +1,5 @@
 import { useEffect } from "react";
 import styled from "styled-components";
-import styles from "../styles/Home.module.css";
 
 export default function Home() {
   useEffect(() => {
