+++ app/styles/Home.module.css
@@ -15,7 +15,7 @@
 }
 
 .main {
-  min-height: 100vh;
+  height: calc(100vh);
   width: 100%;
   /* padding: 4rem 0; */
   flex: 1;
